A hell of a lot of changes
Essentially I tried a lot of ways to include what I have learnt in tkinter into my work and I think its finally gotten to stage that I can make push for it this is the main reason why I haven't commited in a while as I was struggling do to this being my largest hurdle since doing this  project. the new commit I do for this will be cleaning up the files and such as I ended up with 3 different versions of main main which hsa the original main_with tkinkter_v2 which has tkinter elements that I put in the code to help me plan the layout for my new code and new main a new main I added YES I am aware that I could have done this using version control and branches but I didn't want to take any chances. These all come together to help me create tkinter plan which is essentially the main python file NOW THAT USES TKINTER my plan before was to go code code console and then tkinter code console (sounds like gibberish but it is meant for only me and so tkinker plan essentially is the main file I was working with. plan on creationg an alternate folders which contain the task program with tkinter and without tkinter seperately  anyway so here it goes:

I also forgot to say but how I split the files apart for less code in the file I undid that and will redo it once I get my code layout sorted out

another thing is that I only am now doing a push because the program has reached and acceptable point where all of the functions SHOULD be working

- I realised my code was getting to long and tried to split the code into classes functions and main files for easier use this helped in the short term and made it easier to naviagate through my code however it also made it awkward to work with

-all of the functions have now been edited for use with tkinker  and all of the code outside the todolist class has now been added to its corresponding method (asking you what you would like to name the index and stuff all hapens within the class now redicing the layers from 2 to 1)